brought store out of index.ts
